Revert "Revert "OCLOMRS-29: Restructure app to improve code maintainability" (#21)"
This reverts commit f6d6d2839baee51b5929fac090fd04b77968c9d4.
