+++ src/redux/actions/auth/authActionCreators.js
@@ -1,9 +1,4 @@
-
-const AUTHENTICATED = 'loggedin_user';
-const AUTHENTICATION_FAILED = 'login_failed';
-const AUTHENTICATION_IN_PROGRESS = 'login-progess';
-const LOGGED_OUT = 'logout';
-const LOGOUT_FAILED = 'logout_failed';
+import { AUTHENTICATED, AUTHENTICATION_FAILED, AUTHENTICATION_IN_PROGRESS, LOGGED_OUT, LOGOUT_FAILED } from '../types';
 
 // Action creator
 const login = response => ({
@@ -34,11 +29,6 @@ const logoutFailed = errorMessage => ({
 });
 
 export {
-  AUTHENTICATED,
-  AUTHENTICATION_FAILED,
-  AUTHENTICATION_IN_PROGRESS,
-  LOGGED_OUT,
-  LOGOUT_FAILED,
   login,
   loginFailed,
   loginStarted,
