+++ src/redux/actions/types.js
@@ -1,0 +1,5 @@
+export const AUTHENTICATED = '[auth] loggedin user';
+export const AUTHENTICATION_FAILED = '[auth] login failed';
+export const AUTHENTICATION_IN_PROGRESS = '[auth] login progess';
+export const LOGGED_OUT = '[auth] logout';
+export const LOGOUT_FAILED = '[auth] logout failed';
